major refactoring and splitting the code in various scripts
