+++ scripts/utils.py
@@ -1,11 +1,24 @@
 import supervision as sv
 import cv2
 import numpy as np
+import os
+import yaml
+
 
 def resolve_goalkeepers_team_id(
     players: sv.Detections,
     goalkeepers: sv.Detections
 ) -> np.ndarray:
+    """
+    Assign team IDs to goalkeepers based on proximity to team centroids.
+    
+    Args:
+        players: Player detections with team IDs
+        goalkeepers: Goalkeeper detections needing team ID assignment
+        
+    Returns:
+        Array of team IDs for each goalkeeper
+    """
     goalkeepers_xy = goalkeepers.get_anchors_coordinates(sv.Position.BOTTOM_CENTER)
     players_xy = players.get_anchors_coordinates(sv.Position.BOTTOM_CENTER)
     team_0_centroid = players_xy[players.class_id == 0].mean(axis=0)
@@ -17,3 +30,44 @@ def resolve_goalkeepers_team_id(
         goalkeepers_team_id.append(0 if dist_0 < dist_1 else 1)
 
     return np.array(goalkeepers_team_id)
+
+
+def load_config(config_path="config.yaml"):
+    """
+    Load configuration from YAML file.
+    
+    Args:
+        config_path: Path to configuration YAML file
+        
+    Returns:
+        Dictionary containing configuration values
+        
+    Raises:
+        FileNotFoundError: If config file doesn't exist
+        yaml.YAMLError: If config file has invalid YAML
+    """
+    try:
+        with open(config_path, "r") as f:
+            config = yaml.safe_load(f)
+        return config
+    except FileNotFoundError:
+        print(f"Error: Configuration file {config_path} not found.")
+        raise
+    except yaml.YAMLError as e:
+        print(f"Error parsing configuration file: {e}")
+        raise
+
+
+def setup_directories(directories):
+    """
+    Create directories if they don't exist.
+    
+    Args:
+        directories: String or list of directory paths to create
+    """
+    if isinstance(directories, str):
+        directories = [directories]
+        
+    for directory in directories:
+        os.makedirs(directory, exist_ok=True)
+        print(f"Directory created/verified: {directory}")
